moved helper functions from generator & discriminator files and main.py to GANutils.py in keras-analysis-utils
